Refactor create_heredoc
